Update: 挙動を Hook にまとめる、deps (Hooks API) の整理
