+++ src/components/organisms/ChatJoin/component.tsx
@@ -1,45 +1,17 @@
-import React, {
-  Dispatch,
-  SetStateAction,
-  useCallback,
-  useContext,
-  useRef,
-} from "react";
+import React, { Dispatch, SetStateAction } from "react";
 
 // Components
 import { InputForm } from "../../molecules/InputForm";
 
-// Contexts
-import { PeerContext, AllStreamStoreContext } from "../../pages/App";
-import { RoomContext } from "../../pages/App/component";
+// Hooks
+import { useChatJoinHooks } from "./hooks";
 
 type Props = {
   setIsJoinRoom: Dispatch<SetStateAction<boolean>>;
 };
 
 export const ChatJoin: React.VFC<Props> = ({ setIsJoinRoom }) => {
-  const allStreamInfo = useContext(AllStreamStoreContext);
-  const peer = useContext(PeerContext);
-  const { setRoom } = useContext(RoomContext);
-
-  const inputRef = useRef<HTMLInputElement>();
-
-  const onClick = useCallback(() => {
-    // ルーム ID が入力されている場合のみ動作させる
-    const roomId = `${inputRef.current?.value}`;
-
-    if (roomId && peer && allStreamInfo.localStream) {
-      const room = peer.joinRoom(roomId, {
-        mode: "sfu",
-        stream: allStreamInfo.localStream.stream
-          ? allStreamInfo.localStream.stream
-          : undefined,
-      });
-
-      setRoom(room);
-      setIsJoinRoom(true);
-    }
-  }, [peer, allStreamInfo.localStream]);
+  const { inputRef, onClick } = useChatJoinHooks(setIsJoinRoom);
 
   return (
     <InputForm
