+++ src/components/organisms/ChatJoin/index.ts
@@ -1,1 +1,2 @@
 export { ChatJoin } from "./component";
+export { useChatJoinHooks } from "./hooks";
